groth16/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
ape-solidity==0.6.11
asttokens==2.4.1
attrs==23.1.0
base58==1.0.3
bitarray==2.8.3
cached-property==1.5.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
commonmark==0.9.1
cryptography==41.0.7
cytoolz==0.12.2
dataclassy==0.11.1
decorator==5.1.1
Deprecated==1.2.14
eip712==0.2.1
eth-abi==4.2.1
eth-account==0.8.0
eth-ape==0.6.26
eth-bloom==3.0.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-tester==0.9.1b1
eth-typing==3.5.2
eth-utils==2.3.1
ethpm-types==0.5.11
evm-trace==0.1.0a25
executing==2.0.1
frozenlist==1.4.0
galois==0.3.6
hexbytes==0.3.1
idna==3.6
ijson==3.2.3
importlib-metadata==6.8.0
iniconfig==2.0.0
ipython==8.18.1
jedi==0.19.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
lazyasd==0.1.4
llvmlite==0.41.1
lru-dict==1.2.0
matplotlib-inline==0.1.6
morphys==1.0
msgspec==0.18.4
multidict==6.0.4
mypy-extensions==1.0.0
numba==0.58.1
numpy==1.25.2
packaging==23.2
pandas==1.5.3
parsimonious==0.9.0
parso==0.8.3
pexpect==4.9.0
pluggy==1.3.0
prompt-toolkit==3.0.41
protobuf==4.25.1
ptyprocess==0.7.0
pure-eval==0.2.2
py-cid==0.3.0
py-ecc==6.0.0
py-evm==0.7.0a4
py-geth==3.13.0
py-multibase==1.0.3
py-multicodec==0.2.1
py-multihash==0.2.3
py-solc-x==2.0.2
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.5.2
pydantic_core==2.14.5
pyethash==0.1.27
PyGithub==1.59.1
Pygments==2.17.2
PyJWT==2.8.0
PyNaCl==1.5.0
pytest==7.4.3
python-baseconv==1.2.2
python-dateutil==2.8.2
pytz==2023.3.post1
pyunormalize==15.1.0
PyYAML==6.0.1
referencing==0.31.0
regex==2023.10.3
requests==2.31.0
rich==12.6.0
rlp==3.0.0
rpds-py==0.13.1
safe-pysha3==1.0.4
scipy==1.11.4
semantic-version==2.10.0
six==1.16.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.23
stack-data==0.6.3
toolz==0.12.0
tqdm==4.66.1
traitlets==5.14.0
trie==2.2.0
typing_extensions==4.8.0
urllib3==1.26.18
varint==1.0.2
watchdog==3.0.0
wcwidth==0.2.12
web3==6.11.4
websockets==12.0
wrapt==1.16.0
yarl==1.9.3
zipp==3.17.0